# pangea-api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  main.Admin:
    properties:
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      permissions:
        items:
          type: string
        type: array
      role:
        type: string
    type: object
  main.AdminLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  main.AdminLoginResponse:
    properties:
      admin:
        $ref: '#/definitions/main.Admin'
      message:
        type: string
      success:
        type: boolean
      token:
        type: string
      token_info:
        $ref: '#/definitions/main.TokenInfo'
    type: object
...
  main.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  main.LoginResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/main.User'
    type: object
  main.RegisterRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - confirm_password
    - email
    - firstname
    - lastname
    - password
    type: object
  main.RegisterResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      user_id:
        type: string
    type: object
  main.TokenInfo:
    properties:
      expires_at:
        type: integer
      expires_in:
        type: integer
      issued_at:
        type: integer
      refresh_token:
        type: string
      type:
        type: string
    type: object
  main.User:
    properties:
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Pangea UI Backend API with authentication, admin panel, and OAuth integration
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pangea API
  version: "1.0"
paths:
  /admin/login:
    post:
      consumes:
      - application/json
      description: Authenticate admin with email and password
      parameters:
      - description: Admin login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.AdminLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.AdminLoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.AdminLoginResponse'
      summary: Admin login
      tags:
      - admin
...
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout current user and invalidate token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /api/v1/auth/me:
    get:
      consumes:
      - application/json
      description: Get current user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - user
  /api/v1/auth/profile:
    put:
      consumes:
      - application/json
      description: Update current user profile
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - user
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Refresh token
      tags:
      - auth
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: User login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.LoginResponse'
      summary: Login user
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.RegisterResponse'
      summary: Register a new user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"